Better graphs with trace_on. No shapes though.
+ Reused variables create huge graphs: opened issue. Still no solution.
+ Using trace_on instead of keras callback to save a cleaner graph.
  Tensors shapes not shown, however.

+++ data.py
@@ -178,7 +178,7 @@ def load_pair(name_A, name_B, split, **kwargs):
   dataset_B, size_B = load(name_B, split, **kwargs)
 
   # Dataset size is average of the two
-  size = (size_A + size_B) / 2
+  size = int((size_A + size_B) / 2)
 
   return tf.data.Dataset.zip((dataset_A, dataset_B)), size
